Added more Permission controllability:
Added the ability to regulate any Player initiated event in the game
using Permissions, are regulated in the same way as commands except
using the event header in the yml file, event name in the file must
match the event class name for it to work. Also changed a bit of SQL
data handling to prevent table data duplication.